+++ changelog.rst
@@ -3,6 +3,7 @@ v3.0.0
 
 * Drop support for macOS.
 * Enable nowrap mode.
+* Move package from "/etc/magnus" to "/usr/share/magnus"
 
 v2.0.0
 ======